refactor: split ModuleTestingEnvironment into Engines and MainLoop classes
MainLoop has the methods tests tend to interact with,
Engines manages the setup behind the scenes.
